.history/requirements_20250815160433.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# LLM Processing
groq==0.4.1
html2text==2020.1.16
selenium-stealth==1.0.6

# Perplexity Sonar (using requests directly)

# Token counting
tiktoken==0.5.1
